+++ CHANGELOG.md
@@ -6,7 +6,8 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 
 ## Unreleased
 
-None.
+- Support setting visibility of the generated trait directly on the `impl`
+  block. For example: `pub impl i32 { ... }`.
 
 ### Breaking changes